refactoring to cea.dev.build
