+++ cea/interfaces/cli/cea_doc.py
@@ -12,6 +12,7 @@ import importlib
 import cea.config
 import cea.inputlocator
 import cea.scripts
+import dev.build
 from cea import ScriptNotFoundException
 
 __author__ = "Daren Thomas"
@@ -54,7 +55,7 @@ def main(config=None):
 
     script_module = importlib.import_module(cea_script.module)
     try:
-        script_module.main(config)
+        dev.build.main(config)
         print("Execution time: %.2fs" % (datetime.datetime.now() - t0).total_seconds())
     except cea.ConfigError as config_error:
         print('ERROR: %s' % config_error)
